Add files via upload
Changed the format of the page a bit by using flex boxes. The width and height of the columns are 400px so it'll look small for people zoomed out, but the pro of doing this instead of doing `width: 32%` like I was before, is that when people zoom in (or are on mobile devices), the columns don't squish together and the boxes will move themselves to retain their square-ish shape.

Another change is in the middle "projects" section, I'm using a bootstrap carousel so each project can rotate around. I had a lot of trouble resizing the image and making sure my css wasn't "overshadowed" by bootstrap's css. This lead to the ` * {font-family: garamond; color: black;"` line

Will continue to work on this. The carousel text disappears if zoomed in too much. Thinking about using another project instead of the Indonesia visualization one because the picture changes where the carousel arrows are located. Will also work on the "Experience" part, like adding start date and end date periods.

It still needs a lot of work, but I feel like I've leveled up a little in my knowledge of html and css, though it's still low ^^;